Refactor PackagingToolVersions
